# gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: VetManager Pro API
  description: |
    API REST para el sistema veterinario SaaS multinacional VetManager Pro.

    **Autenticación:** Bearer JWT (access token de 15min + refresh token de 7 días)

    **Multi-tenant:** Todos los endpoints filtran datos por `branchId` del token JWT.

    **RBAC:** Cada endpoint requiere un permiso específico según el rol del usuario.
  version: "1.0.0"
  contact:
    name: VetManager Pro
    email: api@vetmanagerpro.com
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.vetmanagerpro.com/api/v1
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /medical-records:
    get:
      tags: [Medical Records]
      summary: Listar historias clínicas
      parameters:
        - in: query
          name: patientId
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string, enum: [open, signed, amended] }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200': { description: Lista de HCs }
    post:
      tags: [Medical Records]
      summary: Abrir nueva historia clínica
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [appointmentId, chiefComplaint]
                  properties:
                    appointmentId:      { type: integer }
                    chiefComplaint:     { type: string }
                    reasonForVisit:     { type: string }
                    visitDate:          { type: string, format: date }
                    notes:              { type: string }
                    weightKg:           { type: number }
                    bodyConditionScore: { type: number }
                    temperatureC:       { type: number }
                - type: object
                  required: [patientId, chiefComplaint]
                  properties:
                    patientId:          { type: integer }
                    chiefComplaint:     { type: string }
                    reasonForVisit:     { type: string }
                    visitDate:          { type: string, format: date }
                    notes:              { type: string }
                    weightKg:           { type: number }
                    bodyConditionScore: { type: number }
                    temperatureC:       { type: number }
      responses:
        '201': { description: HC abierta }
...
  /medical-records/{id}/sign:
    post:
      tags: [Medical Records]
      summary: Firmar / cerrar historia clínica
      description: Valida que tenga anamnesis y al menos un diagnóstico antes de firmar. Una vez firmada no puede editarse.
      responses:
        '200': { description: HC firmada exitosamente }
        '400': { description: HC incompleta - falta anamnesis o diagnóstico }
...
  /medical-records/{id}/diagnoses:
    post:
      tags: [Medical Records]
      summary: Agregar diagnóstico
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [diagnosisName, diagnosisType]
              properties:
                diagnosisName:  { type: string }
                diagnosisType:  { type: string, enum: [definitive,presumptive,differential,rule_out] }
                diagnosisCode:  { type: string, description: "Código CIE o SNOMED" }
                isPrimary:      { type: boolean }
                prognosis:      { type: string, enum: [excellent,good,fair,guarded,poor,grave] }
      responses:
        '201': { description: Diagnóstico agregado }
...
  /medical-records/{id}/prescriptions:
    post:
      tags: [Medical Records]
      summary: Crear receta médica
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required: [medicationName, dose, frequency]
                    properties:
                      medicationId:   { type: integer }
                      medicationName: { type: string }
                      dose:           { type: string }
                      frequency:      { type: string }
                      durationDays:   { type: integer }
      responses:
        '201': { description: Receta creada }

  # ── LAB ────────────────────────────────────────────────────────────────────
...